notes: move sanitization from d_text.rb to note_sanitizer.rb.
